Fix Battletoads & Double dragon crash at first boss.
This game has an inert bug that causes it to read a few values from the
WRAM area at the first boss even though the game lacks WRAM. It needs to
be emulated without WRAM to work (at least if 0xFF is used as the WRAM
initialization value).

Fix by giving all AxROM (mapper 7) games no WRAM for now. This matches
reality too.